renamed struct to proofs and removed unused import
